Move aiohttp-related methods into new 'extensions' file
